Split out EditableFileField handling into own extension
